Funcionalidades quase concluidas. Falta validar e usar repositories

+++ app/Controllers/Http/UsersController.ts
@@ -1,5 +1,7 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
+import Account from 'App/Models/Account'
 import User from 'App/Models/User'
+import { DateTime } from 'luxon'
 
 export default class UsersController {
   public async index({ }: HttpContextContract) {
@@ -8,6 +10,15 @@ export default class UsersController {
 
   public async store({ request, response }: HttpContextContract) {
     let data = request.only(['name', 'email', 'password', 'birthday'])
+
+    let values = data.birthday.split("-")
+    let yearBirthday = values[0]
+    let year = new Date()
+    let actualyear = year.getFullYear()
+
+    if(actualyear - yearBirthday < 18){
+      return response.status(401).send("Apenas maiores de 18 anos podem criar uma conta.")
+    }
 
     let newUser = await User.create(data)
 
@@ -49,10 +60,12 @@ export default class UsersController {
 
     let user = await User.find(id)
 
-    if (!user || auth.user!.id == user.id) {
-      return response.status(404)
+    let account = await Account.query().where('user_id', user!.id)
+    
+    if (!user || auth.user!.id == user.id || account.length > 0) {
+      return response.status(404).send("Não foi possível excluir.")
     }
-
+    
     await user!.delete()
     return response.status(200).send({ message: `Usuário ID ${id} excluído com sucesso!` })
   }
